[feat] add user rating info & merge Frontend/KYR/piece_of_cake into Frontend/POC

See merge request mixkio/e203_common_pjt!104

+++ BackEnd/common_pjt/src/main/java/com/E203/pjt/model/dto/res/UserResVO.java
@@ -20,7 +20,8 @@ public class UserResVO {
     private String userNickname;
     private String userPassword;
     private String userImage;
-    private Integer userRating;
+    private Double userRating;
+//    private Integer userRating;
     private String userLat;
     private String userLng;
     private String userAccount;
